Replace asottile/reorder-python-imports with isort
black >=24 is incompatible with it since it adds a whitespace between
an initial docstring and the import section, which
reorder-python-imports does not like. This is something that is not
likely to be fixed. For reference:

https://github.com/asottile/reorder-python-imports/issues/366
https://github.com/psf/black/issues/4175

This substitution causes some reordering since isort separates
first-party imports from third-party ones by default. We're still
keeping the one line per import format to reduce merge conflicts.