Update Pylint patch version
Renovate should not manage it, at least not minor updates, which can add
new checks breaking CI on generated projects.

3.3 adds a new message `too-many-positional-arguments`, which makes
sense, but produces a ton of lint warnings on probably all existing
modules and can only be fixed by introducing a breaking API change.

Let's stay on the 3.2 train for now and consider how to approach the 3.3
update in the future. Options:

* Raise the allowed positional arguments to a more reasonable value,
  then introduce the API change on functions that would have been called
  via kwargs anyways.
* Just disable the message. This is the likely way forward, but I
  think new projects could benefit from this message, it's just hard to
  transition existing extensions.