Pause CI on pushes to main branch while releasing
This commit causes CI to be skipped on pushes to the main branch after a
release PR has been merged to avoid duplicate runs.

Concurrency groups would pose the risk of cancelling the release
workflow if the pr/push one starts after it. We cannot improve the
event trigger because the push event does not know about the
pull_request one.