Fix CI and dev envs with uv >= 0.6.0 (#172)
`uv<0.6` releases used `requires.txt` from `egg-info` of the sdist as a
fallback. This has been changed. Since Salt requires its legacy
`setup.py`, which in turn requires `setuptools<69`, it could not be
built properly and its dependencies were never installed.

This constrains setuptools for all package builds. In theory, we
could limit this to the installation of Salt in `noxfile.py`, but the
`--build-constraint` parameter requires a file, which would need to be
included or created ad-hoc. Constraining setuptools in this way should
not cause problems in the near future, otherwise we can go this more
complicated way. I might have missed something since I'm not a Python
packaging pro. Better suggestions are welcome.